-Change structs to float, as we are using seconds in float instead of ms as long

+++ src/com/blastedstudios/gdxworld/plugin/mode/animation/AnimationStructRow.java
@@ -65,7 +65,7 @@ public class AnimationStructRow {
 	 * Apply current ui and grab struct
 	 */
 	public AnimationStruct apply(){
-		struct.time = Long.parseLong(timeField.getText());
+		struct.time = Float.parseFloat(timeField.getText());
 		return struct;
 	}
