+++ src/com/blastedstudios/gdxworld/world/animation/AnimationStructTimeSorter.java
@@ -4,6 +4,6 @@ import java.util.Comparator;
 
 class AnimationStructTimeSorter implements Comparator<AnimationStruct>{
 	@Override public int compare(AnimationStruct first, AnimationStruct second) {
-		return ((Long)first.time).compareTo(second.time);
+		return ((Float)first.time).compareTo(second.time);
 	}
 }
